change log level

+++ bin/common.sh
@@ -22,6 +22,7 @@ mklogsdir() {
 }
 
 start() {
+    echo "try to start ${module}"
     getpid ${module}
     if [[ $? -eq 0 ]]; then
         mklogsdir
@@ -69,14 +70,20 @@ stop() {
     if [[ -n ${pid} ]]; then
         echo "killing:
         `ps -p ${pid}`"
-	echo "--------------" ${pid}
-        kill ${pid}
-        if [[ $? -eq 0 ]]; then
-	    rm -rf ./${module}_pid
-            echo "killed"
+	    echo "try to kill" ${pid}
+        pidCount=`ps -p ${pid}|grep ${pid}|wc -l`
+#        `ps -p ${pid}|grep ${pid}|wc -l`
+	    if [[ $pidCount -ne 0 ]]; then
+	        kill ${pid}
+	         if [[ $? -eq 0 ]]; then
+	            rm -rf ./${module}_pid
+                echo "killed"
+            else
+                echo "kill error"
+            fi
         else
-            echo "kill error"
-        fi
+            echo "pid ${pid} is not exist"
+	    fi
     else
         echo "service not running"
     fi
